RBSAN: Introducing "RefBase"-SAN
Inspired by FDSAN, this creates a way to tag binder objects
with ownership information. The first usage is for Java
native object holders, since if there is a double-deletion,
we don't catch it until garbage collection is run.

This is similar to debugRefs, except the implementation is
written to be fast enough that we can always turn it on.

Note: unlike FDSAN, we support multiple different objects
tagging the same binder object. This is because we may have
unrelated code in the same process using a binder object.
For instance, if a service is registered, two libraries
in the same process might both retrieve it and then try
to own it. In this case, we allow the binder to be tagged
with two different owners, but if it's deleted while
either of them are still alive, we through an error. We
currently don't throw an error if there are two declared
owners but only one reference count, but this may be
desirable later.

Bug: 430341301
Test: libutils_binder_test
Flag: EXEMPT bug fix / memory corruption prevention
Change-Id: I502952fa3fc64f4ec53ae747c7df4e18b868d3d3