ashmem: Check if shared ro mappings are broken for write sealed memfds
The kernel previously had a bug where memfds sealed with F_SEAL_WRITE
could not be mapped as read-only and shared [1]. This was clearly a bug
and was fixed on 6.7+. The fixes were also backported to older kernels
but that was done recently, so they have not yet been integrated into
all Android kernels and devices that are being used for development
for Android 17.

Therefore, check if the issue is still present and don't use memfd if
it is. This can be removed once the SDK version is finalized, since
kernels upgrading/launching with Android 17 must run a kernel that
is at least from December 2025, which will have the required fixes.

[1] https://lore.kernel.org/all/cover.1697116581.git.lstoakes@gmail.com/T/#m28fbfb0d5727e5693e54a7fb2e0c9ac30e95eca5
Change-Id: Ia5bb67c0eb2310b00f417e6a968d5fdb06fbf9cc
Bug: 111903542
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>