Refactor USB controller setup to be more robust
Add a new property -- ro.boot.wait_for_udc -- that can be used in
recovery and charger mode to wait for the sys.usb.controller to be
detected before triggering `late-init`. This ensures that fastbootd,
recovery, and charger services don't start before the USB controller is
available which can lead to the USB gadget not being configured
correctly.

If the USB controller never comes up, then we will timeout after 10
seconds to allow init to continue. Init will continue to check if the
USB controller is available in the same way it currently does.

Test: reboot to fastbootd 100x to verify USB detection
Test: reboot to recovery 100x to verify USB detection
Test: reboot to charger mode to verify USB detection
Test: normal reboots x100 and verify adb
Bug: 448889416
Change-Id: I50143ffb7510a6e6a2b3406b1802be83728f8252