Fix debuggerd tests failing on MTE.
Change the executable test to use mmap since this fails on MTE
devices using the local variable.

Remove the LogcatCollector code from tests. This also fails on MTE
on devices and is potentially flaky since it tries to read the
logcat using the shell command logcat. It might be useful
to add a test that verifies that the crash data goes to the logcat,
but it should be done differently to avoid being racy and is probably
not that useful now.

Test: Ran all unit tests on shiba with MTE on and off.
Change-Id: I59efac49bdc35d36d7fcb3e956ba9a781e2f5e77