Remove unnecessary fsync on fingerprint file write
When prefetch runs early in the boot, the device is already very I/O
saturated and we should avoid using unnecessary fsync which further
stall the disk. Given the current script logic, it's not a big deal if
the fingerprint is not recorded properly (this is also very unlikely),
because in the worst case scenario we'll simply re-record the trace in a
later reboot anyway.

Bug: 454451331
Test: CQ
Change-Id: Ief9cbd8850b2f19c22bb71af7860b1d56c446881